retire gcd_brown in dense format
